src/requirements.txt
flask
pandas
sentence-transformers
faiss-cpu
tqdm
openai
python-dotenv
numpy
openpyxl
streamlit
openpyxl
openpyxl

